+++ src/main/java/com/app/dto/PurchasePaymentDTO.java
@@ -1,5 +1,6 @@
 package com.app.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Data;
@@ -8,9 +9,9 @@ public class PurchasePaymentDTO {
 	private LocalDate dateOfPurchase;
     private LocalDate dateOfTransaction;
     private String trans_id;
-    private Double totalAmount;
-    private Double paidAmount;
-    private Double pendingPayment;
+    private BigDecimal totalAmount;
+    private BigDecimal paidAmount;
+    private BigDecimal pendingPayment;
     private String comment;
     private Long supplier;
 }
