+++ src/main/java/com/app/dto/TradingEntryDto.java
@@ -1,5 +1,6 @@
 package com.app.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import lombok.Data;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -14,11 +15,11 @@ public class TradingEntryDto {
     private Long supplierId;
     private Long partyVehicleId;
     private Integer birds;
-    private Double kilograms;
-    private Double rate;
-    private Integer amount;
-    private Integer payment;
-    private Integer pending;
-    private Integer balanceAmount;
+    private BigDecimal kilograms;
+    private BigDecimal rate;
+    private BigDecimal amount;
+    private BigDecimal payment;
+    private BigDecimal pending;
+    private BigDecimal balanceAmount;
     private String description;
 }
